+++ src/main/java/ch/heigvd/sitr/map/Road.java
@@ -1,0 +1,8 @@
+package ch.heigvd.sitr.map;
+
+public class Road {
+
+    Road() {
+        
+    }
+}
