--- src/main/java/ch/heigvd/sitr/map/RoadNetwork.java
@@ -1,10 +1,0 @@
-package ch.heigvd.sitr.map;
-
-public class RoadNetwork {
-
-    private Line line1;
-
-    RoadNetwork() {
-        
-    }
-}
