+++ src/main/java/ch/heigvd/sitr/map/RoadSegment.java
@@ -20,7 +20,7 @@ public class RoadSegment {
     private int yCoordinatesBegin;
     private int xCoordinatesEnd;
     private int yCoordinatesEnd;
-    private LinkedList<Lane> laneList;
+    private LinkedList<LaneSegment> laneList;
 
     RoadSegment(int xCoordinatesBegin, int yCoordinatesBegin, int xCoordinatesEnd, int yCoordinatesEnd,
                 int lenght, float width, int nbLaneLeft, int nbLaneRight, int speedLimit, int roadId) {
@@ -39,12 +39,12 @@ public class RoadSegment {
         float laneWidth = width / (float) (nbLaneLeft + nbLaneRight);
         /*Création des voies sur la gauche de la ligne, id négatif*/
         for (int i = 1; i <= nbLaneLeft; i++) {
-            laneList.add(new Lane(this, xCoordinatesEnd, (yCoordinatesEnd - (int) (laneWidth * i) + (int) (laneWidth / 2)),
+            laneList.add(new LaneSegment(this, xCoordinatesEnd, (yCoordinatesEnd - (int) (laneWidth * i) + (int) (laneWidth / 2)),
                     xCoordinatesBegin, (yCoordinatesBegin - (int) (laneWidth * i) + (int) (laneWidth / 2)), lenght, laneWidth, speedLimit, -i));
         }
         /*Création des voies sur la droite de la ligne, id positif*/
         for (int i = 1; i <= nbLaneRight; i++) {
-            laneList.add(new Lane(this, xCoordinatesBegin, (yCoordinatesBegin + (int) (laneWidth * i) - (int) (laneWidth / 2)),
+            laneList.add(new LaneSegment(this, xCoordinatesBegin, (yCoordinatesBegin + (int) (laneWidth * i) - (int) (laneWidth / 2)),
                     xCoordinatesEnd, (yCoordinatesEnd + (int) (laneWidth * i) - (int) (laneWidth / 2)), lenght, laneWidth, speedLimit, i));
         }
     }
@@ -66,7 +66,7 @@ public class RoadSegment {
 
     /*retourne la ligne sur la gauche de la ligne donnée en paramètre
      * !! Aucun contrôle (pour le moment) si la ligne n'existe pas*/
-    public Lane getLaneOnLeft(int laneId) {
+    public LaneSegment getLaneOnLeft(int laneId) {
         if (laneId < 0) {
             return laneList.get(laneId * (-1));
         } else {
@@ -76,7 +76,7 @@ public class RoadSegment {
 
     /*retourne la ligne sur la droite de la ligne donnée en paramètre
      * !! Aucun contrôle (pour le moment) si la ligne n'existe pas*/
-    public Lane getLaneOnRight(int laneId) {
+    public LaneSegment getLaneOnRight(int laneId) {
         if (laneId < 0) {
             return laneList.get((laneId * (-1)) - 2);
         } else {
@@ -85,7 +85,7 @@ public class RoadSegment {
     }
 
     /*Retourne true si il y a assez de place sur la voie où le véhicule veut se rendre*/
-    public boolean checkIfChangeLaneIsSafe(double relativePosition, Lane lane) {
+    public boolean checkIfChangeLaneIsSafe(double relativePosition, LaneSegment lane) {
         return lane.checkIfChangeLaneIsSafe(relativePosition, lane);
     }
